# specs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Biotech M&A Predictor API
  version: 1.0.0
  description: |
    API for accessing biotech M&A predictions, company intelligence, and market signals.

    This API serves institutional clients (hedge funds, PE firms, investment banks) with:
    - Real-time M&A probability scores and rankings
    - Comprehensive company profiles and drug pipeline data
    - Historical signal tracking and analysis
    - Automated report generation and delivery
    - Configurable alerts for market events

    ## Authentication
    All requests require an API key passed in the `X-API-Key` header.

    ## Rate Limiting
    - Standard tier: 100 requests/minute
    - Premium tier: 500 requests/minute
    - Enterprise tier: Custom limits

    Rate limit information is returned in response headers:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when the window resets

  contact:
    name: API Support
    email: api-support@biotechma.io
  license:
    name: Proprietary
servers:
  - url: https://api.biotechma.io/v1
    description: Production server
  - url: https://api-staging.biotechma.io/v1
    description: Staging server
security:
  - ApiKeyAuth: []
...
paths:
...
  /reports:
    get:
      tags:
        - Reports
      summary: List available reports
      description: |
        Retrieve a list of all available reports including daily digests,
        weekly watchlists, and custom generated reports.
      operationId: listReports
      parameters:
        - name: report_type
          in: query
          description: Filter by report type
          schema:
            type: string
            enum: [daily_digest, weekly_watchlist, custom, signal_alert, deal_analysis]
        - name: start_date
          in: query
          description: Filter reports from this date onwards
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          description: Filter reports up to this date
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Successful response
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReportSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
              example:
                data:
                  - report_id: "rpt_20240115_digest"
                    report_type: "daily_digest"
                    title: "Daily M&A Digest - January 15, 2024"
                    generated_at: "2024-01-15T06:00:00Z"
                    summary: "5 new signals detected, 2 companies moved to high-probability watchlist"
                    download_url: "https://api.biotechma.io/v1/reports/rpt_20240115_digest/download"
                  - report_id: "rpt_20240114_watchlist"
                    report_type: "weekly_watchlist"
                    title: "Weekly M&A Watchlist - Week of January 14, 2024"
                    generated_at: "2024-01-14T06:00:00Z"
                    summary: "25 companies on high-probability watchlist, 8 new target-acquirer pairings identified"
                    download_url: "https://api.biotechma.io/v1/reports/rpt_20240114_watchlist/download"
                pagination:
                  page: 1
                  limit: 20
                  total_results: 145
                  total_pages: 8
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /reports/{report_id}:
    get:
      tags:
        - Reports
      summary: Get specific report
      description: |
        Retrieve the full content of a specific report by its ID.
      operationId: getReport
      parameters:
        - name: report_id
          in: path
          required: true
          description: Unique report identifier
          schema:
            type: string
            example: "rpt_20240115_digest"
        - name: format
          in: query
          description: Report format
          schema:
            type: string
            enum: [json, pdf, html]
            default: json
      responses:
        '200':
          description: Successful response
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
            application/pdf:
              schema:
                type: string
                format: binary
            text/html:
              schema:
                type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /reports/generate:
    post:
      tags:
        - Reports
      summary: Generate on-demand report
      description: |
        Generate a custom report based on specified parameters.
        Report generation is asynchronous; check status via returned report_id.
      operationId: generateReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequest'
            example:
              report_type: "custom"
              title: "Q1 2024 Oncology M&A Analysis"
              parameters:
                therapeutic_areas: ["oncology"]
                min_ma_score: 65.0
                max_market_cap: 2000.0
                include_sections:
                  - "executive_summary"
                  - "top_targets"
                  - "acquirer_analysis"
                  - "deal_predictions"
                format: "pdf"
      responses:
        '202':
          description: Report generation accepted
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                type: object
                properties:
                  report_id:
                    type: string
                    description: Unique identifier for the generated report
                  status:
                    type: string
                    enum: [queued, processing, completed, failed]
                  estimated_completion:
                    type: string
                    format: date-time
                  status_url:
                    type: string
                    description: URL to check report status
              example:
                report_id: "rpt_custom_78945"
                status: "queued"
                estimated_completion: "2024-01-15T15:05:00Z"
                status_url: "https://api.biotechma.io/v1/reports/rpt_custom_78945"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication. Contact support to obtain your key.
  headers:
    X-RateLimit-Limit:
      description: Maximum number of requests allowed in the current time window
      schema:
        type: integer
        example: 100
    X-RateLimit-Remaining:
      description: Number of requests remaining in the current time window
      schema:
        type: integer
        example: 95
    X-RateLimit-Reset:
      description: Unix timestamp when the rate limit window resets
      schema:
        type: integer
        example: 1705329600
  schemas:
...
    ReportSummary:
      type: object
      properties:
        report_id:
          type: string
          example: "rpt_20240115_digest"
        report_type:
          type: string
          enum: [daily_digest, weekly_watchlist, custom, signal_alert, deal_analysis]
        title:
          type: string
          example: "Daily M&A Digest - January 15, 2024"
        generated_at:
          type: string
          format: date-time
        summary:
          type: string
          description: Brief summary of report contents
        download_url:
          type: string
          format: uri
          description: URL to download the report
    Report:
      type: object
      properties:
        report_id:
          type: string
        report_type:
          type: string
        title:
          type: string
        generated_at:
          type: string
          format: date-time
        content:
          type: object
          description: Report content structure varies by report type
          additionalProperties: true
    ReportRequest:
      type: object
      required:
        - report_type
      properties:
        report_type:
          type: string
          enum: [custom, signal_alert, deal_analysis]
        title:
          type: string
          description: Custom report title
        parameters:
          type: object
          description: Report generation parameters
          properties:
            therapeutic_areas:
              type: array
              items:
                type: string
            min_ma_score:
              type: number
              format: float
            max_market_cap:
              type: number
              format: float
            tickers:
              type: array
              items:
                type: string
            include_sections:
              type: array
              items:
                type: string
                enum: [executive_summary, top_targets, acquirer_analysis, deal_predictions, signal_analysis, financial_analysis]
            format:
              type: string
              enum: [json, pdf, html]
              default: pdf
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
          example: 1
        limit:
          type: integer
          description: Results per page
          example: 50
        total_results:
          type: integer
          description: Total number of results
          example: 234
        total_pages:
          type: integer
          description: Total number of pages
          example: 5
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
          example: "invalid_request"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid ticker symbol provided"
        details:
          type: object
          description: Additional error details
          additionalProperties: true
        request_id:
          type: string
          description: Unique request identifier for support
          example: "req_abc123xyz"
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "invalid_request"
            message: "Invalid parameter: min_ma_score must be between 0 and 100"
            request_id: "req_abc123xyz"
    Unauthorized:
      description: Unauthorized - invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "unauthorized"
            message: "Invalid API key provided"
            request_id: "req_abc123xyz"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "not_found"
            message: "Company with ticker 'INVALID' not found"
            request_id: "req_abc123xyz"
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          schema:
            type: integer
            example: 0
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "rate_limit_exceeded"
            message: "Rate limit exceeded. Limit resets in 45 seconds."
            details:
              limit: 100
              window: "1 minute"
              reset_at: 1705329600
            request_id: "req_abc123xyz"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "internal_error"
            message: "An unexpected error occurred. Please try again later."
            request_id: "req_abc123xyz"